Rename `createClassBasedHierarchyDefinition` to `createPredicateBasedHierarchyDefinition`
