+++ packages/hierarchies/learning/Formatting.md
@@ -93,7 +93,7 @@ const hierarchyProvider = createIModelHierarchyProvider({
           // The hierarchy definition returns a single node with a ConcatenatedValue-based label
           {
             node: {
-              key: "custom node",
+              key: { type: "generic", id: "root" },
               label: [
                 "Example | ",
                 {
